refactor: move prisma to shared module
