+++ .gitignore
@@ -11,6 +11,7 @@ coverage
 # next.js
 .next/
 out/
+dist/
 build
 
 # misc
