+++ .npmrc
@@ -1,3 +1,3 @@
-public-hoist-pattern[]=*@emotion*
+public-hoist-pattern[]=*prisma*
 auto-install-peers = true
 node-linker = hoisted
