+++ apps/api/.env.example
@@ -1,13 +1,5 @@
 PORT=8080
 
-POSTGRES_HOST="localhost"
-POSTGRES_PORT="5432"
-POSTGRES_DB="mydb"
-POSTGRES_USER="username"
-POSTGRES_PASSWORD="password"
-
-DATABASE_URL="postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_HOST}:${POSTGRES_PORT}/${POSTGRES_DB}?schema=public"
-
 JWT_SECRET=MyJwtSecret
 JWT_EXPIRES=1h
