--- apps/api/src/common/constants/index.ts
@@ -1,1 +1,0 @@
-export * from './mock'
