+++ apps/api/src/common/mocks/index.ts
@@ -1,2 +1,1 @@
-export * from './getBaseMock'
 export * from './mock'
