+++ apps/api/src/common/mocks/mock/index.ts
@@ -1,6 +1,4 @@
-import { BaseModel, BaseModelType } from 'src/common/types'
-
-import { getBaseMock } from '../getBaseMock'
+import { BaseModel, BaseModelType, getBaseMock } from 'database'
 
 export const mock = <T extends keyof BaseModelType>(model: T) => {
   return new MockConstructor<BaseModelType[T], BaseModelType[T]>(model)
