+++ apps/api/src/database/prisma.service.ts
@@ -1,5 +1,5 @@
 import { INestApplication, Injectable, OnModuleInit } from '@nestjs/common'
-import { PrismaClient } from '@prisma/client'
+import { PrismaClient } from 'database'
 
 @Injectable()
 export class PrismaService extends PrismaClient implements OnModuleInit {
