+++ apps/api/src/modules/auth/auth.repository.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common'
-import { UserType } from '@prisma/client'
+import { UserType } from 'database'
 import { PrismaService } from 'src/database/prisma.service'
 
 import { SignupActorDto, SignupCastingDto } from './auth.dto'
