+++ apps/api/src/modules/auth/auth.service.ts
@@ -6,8 +6,8 @@ import {
 } from '@nestjs/common'
 import { ConfigService } from '@nestjs/config'
 import { JwtService } from '@nestjs/jwt'
-import { User } from '@prisma/client'
 import { compare, hash } from 'bcryptjs'
+import { User } from 'database'
 import { Response } from 'express'
 
 import { LoginDto, SignupActorDto, SignupCastingDto } from './auth.dto'
