+++ apps/api/src/modules/example/example.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty, PartialType } from '@nestjs/swagger'
-import { Example } from '@prisma/client'
 import { IsNotEmpty, IsString } from 'class-validator'
+import { Example } from 'database'
 
 export class GetExampleDto implements Partial<Example> {
   @ApiProperty()
