+++ packages/database/.env.example
@@ -1,0 +1,7 @@
+POSTGRES_HOST="localhost"
+POSTGRES_PORT="5432"
+POSTGRES_DB="mydb"
+POSTGRES_USER="username"
+POSTGRES_PASSWORD="password"
+
+DATABASE_URL="postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_HOST}:${POSTGRES_PORT}/${POSTGRES_DB}?schema=public"
