+++ packages/database/src/constants.ts
