+++ packages/database/src/mocks/index.ts
@@ -1,0 +1,1 @@
+export * from './getBaseMock'
